pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "condition-monitoring-rul"
version = "1.0.0"
description = "Advanced Condition Monitoring and Remaining Useful Life Prediction"
authors = [
    {name = "mwasifanwar", email = "wasifsdk@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
requires-python = ">=3.8"
dependencies = [
    "tensorflow>=2.12.0",
    "scikit-learn>=1.2.0",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "plotly>=5.13.0",
    "scipy>=1.9.0",
    "PyYAML>=6.0",
    "tqdm>=4.64.0",
    "mlflow>=2.4.0",
    "optuna>=3.1.0",
    "joblib>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "flake8>=5.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  | \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''